fix(derive): providers refactor and trait dedup
