+++ crates/derive/src/online/alloy_providers.rs
@@ -1,12 +1,8 @@
 //! This module contains concrete implementations of the data provider traits, using an alloy
 //! provider on the backend.
 
-use crate::{
-    traits::{ChainProvider, L2ChainProvider},
-    types::{
-        Block, BlockInfo, L2BlockInfo, L2ExecutionPayloadEnvelope, OpBlock, RollupConfig,
-        SystemConfig,
-    },
+use crate::types::{
+    Block, BlockInfo, L2BlockInfo, L2ExecutionPayloadEnvelope, OpBlock, RollupConfig, SystemConfig,
 };
 use alloc::{boxed::Box, sync::Arc, vec::Vec};
 use alloy_consensus::{Header, Receipt, ReceiptWithBloom, TxEnvelope, TxType};
@@ -17,6 +13,7 @@ use alloy_transport_http::Http;
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use core::num::NonZeroUsize;
+use kona_providers::{ChainProvider, L2ChainProvider};
 use lru::LruCache;
 
 const CACHE_SIZE: usize = 16;
