+++ crates/derive/src/sources/calldata.rs
@@ -1,13 +1,14 @@
 //! CallData Source
 
 use crate::{
-    traits::{AsyncIterator, ChainProvider, SignedRecoverable},
+    traits::{AsyncIterator, SignedRecoverable},
     types::{BlockInfo, StageError, StageResult},
 };
 use alloc::{boxed::Box, collections::VecDeque};
 use alloy_consensus::{Transaction, TxEnvelope};
 use alloy_primitives::{Address, Bytes, TxKind};
 use async_trait::async_trait;
+use kona_providers::ChainProvider;
 
 /// A data iterator that reads from calldata.
 #[derive(Debug, Clone)]
