+++ crates/derive/src/sources/factory.rs
@@ -2,29 +2,27 @@
 
 use crate::{
     sources::{BlobSource, CalldataSource, DataSource, PlasmaSource},
-    traits::{BlobProvider, ChainProvider, DataAvailabilityProvider},
+    traits::{BlobProvider, DataAvailabilityProvider},
     types::{BlockID, BlockInfo, RollupConfig},
 };
 use alloc::{boxed::Box, fmt::Debug};
 use alloy_primitives::{Address, Bytes};
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
-use kona_plasma::traits::{ChainProvider as PlasmaChainProvider, PlasmaInputFetcher};
+use kona_plasma::traits::PlasmaInputFetcher;
+use kona_providers::ChainProvider;
 
 /// A factory for creating a calldata and blob provider.
 #[derive(Debug, Clone, Copy)]
-pub struct DataSourceFactory<C, B, PCP, PIF, I>
+pub struct DataSourceFactory<C, B, PIF, I>
 where
-    C: ChainProvider + Clone,
+    C: ChainProvider + Send + Clone,
     B: BlobProvider + Clone,
-    PCP: PlasmaChainProvider + Send + Clone,
-    PIF: PlasmaInputFetcher<PCP> + Clone,
+    PIF: PlasmaInputFetcher<C> + Clone,
     I: Iterator<Item = Bytes> + Send + Clone,
 {
     /// The chain provider to use for the factory.
     pub chain_provider: C,
-    /// The plasma chain provider.
-    pub plasma_chain_provider: PCP,
     /// The plasma iterator.
     pub plasma_source: I,
     /// The blob provider
@@ -39,19 +37,17 @@ where
     pub signer: Address,
 }
 
-impl<C, B, PCP, PIF, I> DataSourceFactory<C, B, PCP, PIF, I>
+impl<C, B, PIF, I> DataSourceFactory<C, B, PIF, I>
 where
-    C: ChainProvider + Clone + Debug,
+    C: ChainProvider + Send + Clone + Debug,
     B: BlobProvider + Clone + Debug,
-    PCP: PlasmaChainProvider + Send + Clone + Debug,
-    PIF: PlasmaInputFetcher<PCP> + Clone + Debug,
+    PIF: PlasmaInputFetcher<C> + Clone + Debug,
     I: Iterator<Item = Bytes> + Send + Clone,
 {
     /// Creates a new factory.
-    pub fn new(provider: C, blobs: B, pcp: PCP, pif: PIF, s: I, cfg: &RollupConfig) -> Self {
+    pub fn new(provider: C, blobs: B, pif: PIF, s: I, cfg: &RollupConfig) -> Self {
         Self {
             chain_provider: provider,
-            plasma_chain_provider: pcp,
             plasma_source: s,
             blob_provider: blobs,
             plasma_input_fetcher: pif,
@@ -63,16 +59,15 @@ where
 }
 
 #[async_trait]
-impl<C, B, PCP, PIF, I> DataAvailabilityProvider for DataSourceFactory<C, B, PCP, PIF, I>
+impl<C, B, PIF, I> DataAvailabilityProvider for DataSourceFactory<C, B, PIF, I>
 where
     C: ChainProvider + Send + Sync + Clone + Debug,
     B: BlobProvider + Send + Sync + Clone + Debug,
-    PCP: PlasmaChainProvider + Send + Sync + Clone + Debug,
-    PIF: PlasmaInputFetcher<PCP> + Send + Sync + Clone + Debug,
+    PIF: PlasmaInputFetcher<C> + Send + Sync + Clone + Debug,
     I: Iterator<Item = Bytes> + Send + Sync + Clone + Debug,
 {
     type Item = Bytes;
-    type DataIter = DataSource<C, B, PCP, PIF, I>;
+    type DataIter = DataSource<C, B, PIF, I>;
 
     async fn open_data(
         &self,
@@ -102,7 +97,7 @@ where
         } else if self.plasma_enabled {
             let id = BlockID { hash: block_ref.hash, number: block_ref.number };
             Ok(DataSource::Plasma(PlasmaSource::new(
-                self.plasma_chain_provider.clone(),
+                self.chain_provider.clone(),
                 self.plasma_input_fetcher.clone(),
                 self.plasma_source.clone(),
                 id,
