+++ crates/derive/src/sources/source.rs
@@ -2,22 +2,22 @@
 
 use crate::{
     sources::{BlobSource, CalldataSource, PlasmaSource},
-    traits::{AsyncIterator, BlobProvider, ChainProvider},
+    traits::{AsyncIterator, BlobProvider},
     types::StageResult,
 };
 use alloc::boxed::Box;
 use alloy_primitives::Bytes;
 use async_trait::async_trait;
-use kona_plasma::traits::{ChainProvider as PlasmaChainProvider, PlasmaInputFetcher};
+use kona_plasma::traits::PlasmaInputFetcher;
+use kona_providers::ChainProvider;
 
 /// An enum over the various data sources.
 #[derive(Debug, Clone)]
-pub enum DataSource<CP, B, PCP, PIF, I>
+pub enum DataSource<CP, B, PIF, I>
 where
     CP: ChainProvider + Send,
     B: BlobProvider + Send,
-    PCP: PlasmaChainProvider + Send,
-    PIF: PlasmaInputFetcher<PCP> + Send,
+    PIF: PlasmaInputFetcher<CP> + Send,
     I: Iterator<Item = Bytes> + Send,
 {
     /// A calldata source.
@@ -25,16 +25,15 @@ where
     /// A blob source.
     Blob(BlobSource<CP, B>),
     /// A plasma source.
-    Plasma(PlasmaSource<PCP, PIF, I>),
+    Plasma(PlasmaSource<CP, PIF, I>),
 }
 
 #[async_trait]
-impl<CP, B, PCP, PIF, I> AsyncIterator for DataSource<CP, B, PCP, PIF, I>
+impl<CP, B, PIF, I> AsyncIterator for DataSource<CP, B, PIF, I>
 where
     CP: ChainProvider + Send,
     B: BlobProvider + Send,
-    PCP: PlasmaChainProvider + Send,
-    PIF: PlasmaInputFetcher<PCP> + Send,
+    PIF: PlasmaInputFetcher<CP> + Send,
     I: Iterator<Item = Bytes> + Send,
 {
     type Item = Bytes;
