+++ crates/plasma/src/traits.rs
@@ -1,35 +1,14 @@
 //! Traits for plasma sources and internal components.
 
 use crate::types::{FinalizedHeadSignal, PlasmaError};
-use alloc::{boxed::Box, vec::Vec};
-use alloy_consensus::{Header, Receipt, TxEnvelope};
-use alloy_primitives::{Bytes, B256};
+use alloc::boxed::Box;
+use alloy_primitives::Bytes;
 use async_trait::async_trait;
 use kona_primitives::{
     block::{BlockID, BlockInfo},
     system_config::SystemConfig,
 };
-
-/// Describes the functionality of a data source that can provide information from the blockchain.
-#[async_trait]
-pub trait ChainProvider {
-    /// Fetch the L1 [Header] for the given [B256] hash.
-    async fn header_by_hash(&mut self, hash: B256) -> anyhow::Result<Header>;
-
-    /// Returns the block at the given number, or an error if the block does not exist in the data
-    /// source.
-    async fn block_info_by_number(&mut self, number: u64) -> anyhow::Result<BlockInfo>;
-
-    /// Returns all receipts in the block with the given hash, or an error if the block does not
-    /// exist in the data source.
-    async fn receipts_by_hash(&mut self, hash: B256) -> anyhow::Result<Vec<Receipt>>;
-
-    /// Returns the [BlockInfo] and list of [TxEnvelope]s from the given block hash.
-    async fn block_info_and_transactions_by_hash(
-        &mut self,
-        hash: B256,
-    ) -> anyhow::Result<(BlockInfo, Vec<TxEnvelope>)>;
-}
+use kona_providers::ChainProvider;
 
 /// A plasma input fetcher.
 #[async_trait]
