+++ crates/providers/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "kona-providers"
+description = "Provider traits and implementations for kona crates"
+version = "0.0.1"
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+tracing.workspace = true
+alloy-rlp = { workspace = true, features = ["derive"] }
+alloy-primitives = { workspace = true, features = ["rlp"] }
+
+# Local
+kona-primitives = { path = "../primitives", version = "0.0.1" }
+
+# External
+async-trait = "0.1.77"
+hashbrown = { version = "0.14.3", optional = true }
+
+# Alloy Types
+alloy-sol-types = { version = "0.7.1", default-features = false }
+alloy-consensus = { git = "https://github.com/alloy-rs/alloy", rev = "e3f2f07", default-features = false }
+op-alloy-consensus = { git = "https://github.com/clabby/op-alloy", branch = "refcell/consensus-port", default-features = false }
+
+# `serde` feature dependencies
+serde = { version = "1.0.197", default-features = false, features = ["derive"], optional = true }
+
+[dev-dependencies]
+hashbrown = "0.14.3"
+tokio = { version = "1.37", features = ["full"] }
+tracing-subscriber = "0.3.18"
+serde_json = { version = "1.0.68", default-features = false }
+
+[features]
+default = ["serde"]
+serde = ["dep:serde"]
+test-utils = [ "dep:hashbrown" ]
